move enum valtype to a separate header
replace forward decls of the enum with inclusions of the header.
forward decl of an enum doesn't make much sense as the size of
incomplete enum is in general unknown.
